[fuse_ctrl, test] Separate fuse_ctrl and lc_ctrl libs
