+++ src/integration/test_suites/libs/lc_ctrl/lc_ctrl.h
@@ -1,0 +1,31 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright 2020 Western Digital Corporation or its affiliates.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+#ifndef LC_CTRL_H
+#define LC_CTRL_H
+
+#include <stdint.h>
+
+void lcc_initialization(void);
+
+void sw_transition_req(uint32_t next_lc_state,
+                        uint32_t token_127_96,
+                        uint32_t token_95_64,
+                        uint32_t token_63_32,
+                        uint32_t token_31_0,
+                        uint32_t conditional);
+
+#endif // LC_CTRL_H
